backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
httpx==0.25.2
python-dotenv==1.0.0
chromadb==0.4.18
sentence-transformers==2.2.2
ollama==0.1.7
atlassian-python-api==3.41.0
requests==2.31.0
slack-sdk==3.26.1
click==8.1.7
rich==13.7.0
oci==2.119.1
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography>=41.0.0